Usar un byte para direccion, agregar bit para control de avance/reversa

+++ tablero/src/controls/includes/controls.h
@@ -2,16 +2,16 @@
  * controls.h
  *
  *  Created on: Nov 18, 2023
- *      Author: f3rm3
+ *      Author: Fernando Mendoza
  */
 
 #ifndef SRC_CONTROLS_INCLUDES_CONTROLS_H_
 #define SRC_CONTROLS_INCLUDES_CONTROLS_H_
 
-#include "../../adc10/include/adc10.h"
+#include "adc10.h"
 
 void controls_init(void);
 
-uint8_t controls_get_state(uint8_t * out_speed, uint16_t * out_direction);
+uint8_t controls_get_state(uint8_t * out_speed, uint8_t * out_direction);
 
 #endif /* SRC_CONTROLS_INCLUDES_CONTROLS_H_ */
